修正头文件Rect的定义

+++ src/base_algorithm.h
@@ -12,7 +12,7 @@
 namespace cv
 {
     struct Rect {
-        float x, y, width, height;
+        int x, y, width, height;
     };
 }
 #endif
